axiam_api_grpc::proto::authorization_service_server

Trait AuthorizationService

pub trait AuthorizationService:
    Send
    + Sync
    + 'static {
    // Required methods
    fn check_access<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CheckAccessRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CheckAccessResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn batch_check_access<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BatchCheckAccessRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BatchCheckAccessResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with AuthorizationServiceServer.

Required Methods

fn check_access<'life0, 'async_trait>( &'life0 self, request: Request<CheckAccessRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CheckAccessResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check whether a subject can perform an action on a resource.

fn batch_check_access<'life0, 'async_trait>( &'life0 self, request: Request<BatchCheckAccessRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BatchCheckAccessResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Check multiple access requests in a single round-trip.